Adding tests to check if our query is finding entities without any components
We verified that the entity map and the query map are correct when querying for entities. The possible bug was that entities could be queried for that didn't have any components. However we couldn't reproduce the bug with a test. We then used the VS Code debugger to step through the function and verified that the logic is running correctly. I'm leaving the tests in just in case I'm missing something.